pyproject.toml
[tool.poetry]
name = "resource_tracking"
version = "1.4.21"
description = "DBCA internal corporate application to download and serve data from remote tracking devices."
authors = ["DBCA OIM <asi@dbca.wa.gov.au>"]
license = "Apache-2.0"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "4.2.17"
psycopg = { version = "3.2.3", extras = ["binary", "pool"] }
dbca-utils = "2.0.2"
python-dotenv = "1.0.1"
dj-database-url = "2.3.0"
gunicorn = "23.0.0"
django-extensions = "3.2.3"
django-tastypie = "0.14.7"
django-geojson = "4.1.0"
unicodecsv = "0.14.1"
whitenoise = { version = "6.8.2", extras = ["brotli"] }
azure-storage-blob = "12.23.1"
sentry-sdk = { version = "2.19.0", extras = ["django"] }
uvicorn = {extras = ["standard"], version = "^0.34.0"}
uvicorn-worker = "^0.3.0"
orjson = "3.10.13"

[tool.poetry.group.dev.dependencies]
ipython = "^8.31.0"
ipdb = "^0.13.13"
pre-commit = "^4.0.1"
mixer = "^7.2.2"

# Reference: https://docs.astral.sh/ruff/configuration/
[tool.ruff]
line-length = 120

[tool.ruff.lint]
ignore = [
  "E501", # Line too long
  "E722", # Bare except
]

# Reference: https://www.djlint.com/docs/configuration/
[tool.djlint]
profile = "django"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.masonry.api"
